[nsinsider] Make setup helpers for nsinsider reusable
